requirements.txt
asgiref==3.8.1
build==1.2.2.post1
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
coinbase-commerce==1.0.1
colorama==0.4.6
cryptography==46.0.3
Django==5.2.1
django-appconf==1.2.0
django-dbbackup==5.0.1
django-extensions==4.1
django-imagekit==6.0.0
django-widget-tweaks==1.5.0
git-filter-repo==2.47.0
idna==3.10
MarkupSafe==3.0.3
packaging==25.0
pilkit==3.0
pillow==11.2.1
pip-tools==7.4.1
pony==0.7.19
pycparser==2.23
pyOpenSSL==25.3.0
pyproject_hooks==1.2.0
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
SHop==0.0.2
signals==0.0.2
six==1.17.0
sqlparse==0.5.3
store==2020.2.10
stripe==12.3.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
